+++ PSGSuite/Public/Data Transfer/Start-GSDataTransfer.ps1
@@ -2,48 +2,47 @@
     <#
     .SYNOPSIS
     Starts a Data Transfer from one user to another
-    
+
     .DESCRIPTION
     Starts a Data Transfer from one user to another
-    
+
     .PARAMETER OldOwnerUserId
     The email or unique Id of the owner you are transferring data *FROM*
-    
+
     .PARAMETER NewOwnerUserId
     The email or unique Id of the owner you are transferring data *TO*
-    
+
     .PARAMETER ApplicationId
     The application Id that you would like to transfer data for
-    
+
     .PARAMETER PrivacyLevel
     The privacy level for the data you'd like to transfer
 
     Available values are:
     * "SHARED": all shared content owned by the user
     * "PRIVATE": all private (unshared) content owned by the user
-    
+
     .EXAMPLE
     Start-GSDataTransfer -OldOwnerUserId joe -NewOwnerUserId mark -ApplicationId 55656082996 -PrivacyLevel SHARED,PRIVATE
 
     Transfers all of Joe's data to Mark
     #>
     [cmdletbinding()]
-    Param
-    (
-      [parameter(Mandatory=$true,Position=0)]
-      [string]
-      $OldOwnerUserId,
-      [parameter(Mandatory=$true,Position=1)]
-      [string]
-      $NewOwnerUserId,
-      [parameter(Mandatory=$true,Position=2,ValueFromPipelineByPropertyName=$true)]
-      [alias("id")]
-      [string]
-      $ApplicationId,
-      [parameter(Mandatory=$true)]
-      [ValidateSet("SHARED","PRIVATE")]
-      [string[]]
-      $PrivacyLevel
+    Param (
+        [parameter(Mandatory=$true,Position=0)]
+        [string]
+        $OldOwnerUserId,
+        [parameter(Mandatory=$true,Position=1)]
+        [string]
+        $NewOwnerUserId,
+        [parameter(Mandatory=$true,Position=2,ValueFromPipelineByPropertyName=$true)]
+        [alias("id")]
+        [string]
+        $ApplicationId,
+        [parameter(Mandatory=$true)]
+        [ValidateSet("SHARED","PRIVATE")]
+        [string[]]
+        $PrivacyLevel
     )
     Begin {
         $serviceParams = @{