+++ Tests/Module Tests/PSGSuite.Tests.ps1
@@ -1,7 +1,7 @@
 $PSVersion = $PSVersionTable.PSVersion.Major
-$ModuleName = "PSGSuite"
-$projectRoot = Resolve-Path "$PSScriptRoot\.."
-$ModulePath = Resolve-Path "$projectRoot\$ModuleName"
+$projectRoot = Resolve-Path "$PSScriptRoot\..\.."
+$ModulePath = Resolve-Path "$projectRoot\BuildOutput\$($env:BHProjectName)"
+$decompiledModulePath = Resolve-Path "$projectRoot\$($env:BHProjectName)"
 $env:EnablePSGSuiteDebug = $true
 
 # Verbose output for non-master builds on appveyor
@@ -12,29 +12,15 @@ if ($ENV:BHBranchName -eq "development" -or $env:BHCommitMessage -match "!verbos
     $Verbose.add("Verbose",$True)
 }
 
-$moduleRoot = Split-Path (Resolve-Path "$projectRoot\*\*.psd1")
+$moduleRoot = Split-Path (Resolve-Path "$ModulePath\*\*.psd1")
 
-Import-Module 'Configuration' -RequiredVersion 1.2.0
-Import-Module $ModulePath -Force
+Import-Module $ModulePath -Force -Verbose:$false
 
-Describe "Previous build validation" {
-    Context "Failure breadcrumb from previous build" {
-        It "Should not exist" {
-            "$projectRoot\BuildFailed.txt" | Should -Not -Exist
-        }
-    }
-}
-
-Describe "Module tests: $ModuleName" {
-    Context "Confirm private functions are not exported on module import" {
-        It "Should throw when checking for Get-GSUserListPrivate in the exported commands" {
-            {Get-Command -Name Get-GSUserListPrivate -Module PSGSuite -ErrorAction Stop} | Should -Throw "The term 'Get-GSUserListPrivate' is not recognized as the name of a cmdlet, function, script file, or operable program."
-        }
-    }
+Describe "Module tests: $($env:BHProjectName)" {
     Context "Confirm files are valid Powershell syntax" {
-        $scripts = Get-ChildItem $ModulePath -Include *.ps1,*.psm1,*.psd1 -Recurse
+        $scripts = Get-ChildItem $decompiledModulePath -Include *.ps1,*.psm1,*.psd1 -Recurse
 
-        $testCase = $scripts | Foreach-Object {@{file = $_}}         
+        $testCase = $scripts | Foreach-Object {@{file = $_}}
         It "Script <file> should be valid Powershell" -TestCases $testCase {
             param($file)
 
@@ -46,8 +32,15 @@ Describe "Module tests: $ModuleName" {
             $errors.Count | Should Be 0
         }
     }
+    Context "Confirm private functions are not exported on module import" {
+        $testCase = Get-ChildItem "$decompiledModulePath\Private" -Recurse -Include *.ps1 | Foreach-Object {@{item = $_.BaseName}}
+        It "Should throw when checking for '<item>' in the module commands" -TestCases $testCase {
+            param($item)
+            {Get-Command -Name $item -Module $env:BHProjectName -ErrorAction Stop} | Should -Throw "The term '$item' is not recognized as the name of a cmdlet, function, script file, or operable program."
+        }
+    }
     Context "Confirm all aliases are created" {
-        $aliasHash = . "$ModulePath\Aliases\PSGSuite.Aliases.ps1"
+        $aliasHash = . "$decompiledModulePath\Aliases\PSGSuite.Aliases.ps1"
 
         $testCase = $aliasHash.Keys | ForEach-Object {@{Name = $_;Value = $aliasHash[$_]}}
 
@@ -60,8 +53,8 @@ Describe "Module tests: $ModuleName" {
     }
     Context "Confirm there are no duplicate function names in private and public folders" {
         It 'Should have no duplicate functions' {
-            $functions = Get-ChildItem "$moduleRoot\Public" -Recurse -Include *.ps1 | Select-Object -ExpandProperty BaseName
-            $functions += Get-ChildItem "$moduleRoot\Private" -Recurse -Include *.ps1 | Select-Object -ExpandProperty BaseName
+            $functions = Get-ChildItem "$decompiledModulePath\Public" -Recurse -Include *.ps1 | Select-Object -ExpandProperty BaseName
+            $functions += Get-ChildItem "$decompiledModulePath\Private" -Recurse -Include *.ps1 | Select-Object -ExpandProperty BaseName
             ($functions | Group-Object | Where-Object {$_.Count -gt 1}).Count | Should -BeLessThan 1
         }
     }
@@ -69,15 +62,15 @@ Describe "Module tests: $ModuleName" {
 
 Describe "Function contents" {
     Context "All non-helper public functions should use Write-Verbose" {
-        $scripts = Get-ChildItem "$ModulePath\Public" -Include *.ps1 -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
-        $testCase = $scripts | Foreach-Object {@{file = $_;Name = $_.BaseName}}         
+        $scripts = Get-ChildItem "$decompiledModulePath\Public" -Include *.ps1 -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
+        $testCase = $scripts | Foreach-Object {@{file = $_;Name = $_.BaseName}}
         It "Function <Name> should contain verbose output" -TestCases $testCase {
             param($file,$Name)
             $file.fullname | Should -FileContentMatch 'Write-Verbose'
         }
     }
     Context "All 'Remove' functions should SupportsShouldProcess" {
-        $scripts = Get-ChildItem "$ModulePath\Public" -Include 'Remove-*.ps1' -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
+        $scripts = Get-ChildItem "$decompiledModulePath\Public" -Include 'Remove-*.ps1' -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
         $testCase = $scripts | Foreach-Object {@{file = $_;Name = $_.BaseName}}
         It "Function <Name> should contain SupportsShouldProcess" -TestCases $testCase {
             param($file,$Name)
@@ -85,8 +78,8 @@ Describe "Function contents" {
         }
     }
     Context "All 'Remove' functions should contain 'PSCmdlet.ShouldProcess'" {
-        $scripts = Get-ChildItem "$ModulePath\Public" -Include 'Remove-*.ps1' -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
-        $testCase = $scripts | Foreach-Object {@{file = $_;Name = $_.BaseName}}    
+        $scripts = Get-ChildItem "$decompiledModulePath\Public" -Include 'Remove-*.ps1' -Recurse | Where-Object {$_.FullName -notlike "*Helpers*"}
+        $testCase = $scripts | Foreach-Object {@{file = $_;Name = $_.BaseName}}
         It "Function <Name> should contain PSCmdlet.ShouldProcess" -TestCases $testCase {
             param($file,$Name)
             $file.fullname | Should -FileContentMatch 'PSCmdlet.ShouldProcess'